refactor(lint): share access-guard analysis between missing-events lints
The missing-events-access-control and missing-events-arithmetic lints each
carried their own copy of the caller-check detection (guard search, sender
tracking, name heuristic and the statement walkers behind them). Move a single
implementation to `analysis/access_control.rs`, keeping the stronger half of
each copy: sender-alias tracking, guards inside `require(...)` arguments and
nested call expressions, the prefix-before-placeholder cut and the wider name
list from the access-control lint; the check polarity and dominating-statement
restriction from the arithmetic lint. The name heuristic now also requires the
callee to return nothing so a helper whose boolean result is discarded is not
mistaken for a guard. Generic statement walkers move to `analysis/stmts.rs`.
