+++ crates/lint/src/sol/analysis/mod.rs
@@ -1,11 +1,13 @@
 //! Shared, side-effect-free HIR probes reused by Solidity lints.
 
+mod access_control;
 mod exprs;
 mod helper_cache;
 mod modifier_outcome;
 mod stmts;
 mod types;
 
+pub use access_control::*;
 pub use exprs::*;
 pub use helper_cache::*;
 pub use modifier_outcome::*;
