+++ crates/lint/src/sol/low/missing_events_access_control.rs
@@ -4,14 +4,14 @@ use crate::{
     sol::{
         Severity, SolLint,
         analysis::{
-            branch_always_exits, for_each_lhs_var, function_ids, is_builtin, is_msg_sender,
-            is_require_or_assert, lhs_local_var, referenced_item, underlying_var,
+            branch_always_exits, for_each_lhs_var, function_ids, guard_vars, is_protected,
+            is_sender_member, lhs_local_var, referenced_item, underlying_var,
         },
     },
 };
 use solar::{
     ast::{ContractKind, DataLocation, LitKind, StateMutability, Visibility},
-    interface::{Span, data_structures::Never, kw, sym},
+    interface::{Span, data_structures::Never},
     sema::hir::{
         self, EventId, Expr, ExprKind, FunctionId, ItemId, Stmt, StmtKind, VariableId, Visit,
     },
@@ -95,61 +95,6 @@ impl<'hir> LateLintPass<'hir> for MissingEventsAccessControl {
     }
 }
 
-// --- Generic HIR walking ----------------------------------------------------------------------
-
-/// Runs `f` on every statement (pre-order, nested ones included) until it breaks.
-struct StmtVisitor<'hir, F> {
-    hir: &'hir hir::Hir<'hir>,
-    f: F,
-}
-
-impl<'hir, F: FnMut(&'hir Stmt<'hir>) -> ControlFlow<()>> Visit<'hir> for StmtVisitor<'hir, F> {
-    type BreakValue = ();
-
-    fn hir(&self) -> &'hir hir::Hir<'hir> {
-        self.hir
-    }
-
-    fn visit_stmt(&mut self, stmt: &'hir Stmt<'hir>) -> ControlFlow<()> {
-        (self.f)(stmt)?;
-        self.walk_stmt(stmt)
-    }
-}
-
-fn visit_stmts<'hir>(
-    hir: &'hir hir::Hir<'hir>,
-    stmts: &'hir [Stmt<'hir>],
-    f: impl FnMut(&'hir Stmt<'hir>) -> ControlFlow<()>,
-) -> ControlFlow<()> {
-    let mut visitor = StmtVisitor { hir, f };
-    stmts.iter().try_for_each(|stmt| visitor.visit_stmt(stmt))
-}
-
-/// The expression directly owned by `stmt` (nested statements excluded).
-fn stmt_expr<'hir>(hir: &'hir hir::Hir<'hir>, stmt: &'hir Stmt<'hir>) -> Option<&'hir Expr<'hir>> {
-    match stmt.kind {
-        StmtKind::DeclSingle(var_id) => hir.variable(var_id).initializer,
-        StmtKind::DeclMulti(_, expr)
-        | StmtKind::Expr(expr)
-        | StmtKind::Emit(expr)
-        | StmtKind::Revert(expr)
-        | StmtKind::Return(Some(expr))
-        | StmtKind::If(expr, ..) => Some(expr),
-        StmtKind::Try(try_stmt) => Some(&try_stmt.expr),
-        _ => None,
-    }
-}
-
-/// The statements of `body` that run before the `_` placeholder (all of them for functions).
-fn prefix_stmts<'hir>(body: hir::Block<'hir>) -> &'hir [Stmt<'hir>] {
-    let end = body
-        .stmts
-        .iter()
-        .position(|stmt| matches!(stmt.kind, StmtKind::Placeholder))
-        .unwrap_or(body.stmts.len());
-    &body.stmts[..end]
-}
-
 /// Calls `f` on every index and slice bound along the spine of an lvalue.
 fn for_each_lhs_index<'hir>(expr: &'hir Expr<'hir>, f: &mut impl FnMut(&'hir Expr<'hir>)) {
     match &expr.peel_parens().kind {
@@ -171,245 +116,6 @@ fn for_each_lhs_index<'hir>(expr: &'hir Expr<'hir>, f: &mut impl FnMut(&'hir Exp
         ExprKind::Tuple(exprs) => exprs.iter().flatten().for_each(|e| for_each_lhs_index(e, f)),
         _ => {}
     }
-}
-
-// --- Access-control detection -----------------------------------------------------------------
-
-/// `msg.sender` or `tx.origin`.
-fn is_sender_member(expr: &Expr<'_>) -> bool {
-    is_msg_sender(expr)
-        || matches!(&expr.peel_parens().kind, ExprKind::Member(base, name)
-            if name.name == kw::Origin && is_builtin(base, sym::tx))
-}
-
-fn looks_like_access_control(func: &hir::Function<'_>) -> bool {
-    let Some(name) = func.name else { return false };
-    let lower = name.as_str().to_ascii_lowercase();
-    matches!(lower.as_str(), "auth" | "requiresauth" | "restricted")
-        || ["only", "check", "_check"].iter().any(|prefix| {
-            ["admin", "guardian", "manager", "owner", "role"]
-                .iter()
-                .any(|role| lower.starts_with(&format!("{prefix}{role}")))
-        })
-}
-
-/// Locals initialized or assigned from a value that reads `msg.sender`.
-fn sender_aliases<'hir>(
-    hir: &'hir hir::Hir<'hir>,
-    stmts: &'hir [Stmt<'hir>],
-) -> HashSet<VariableId> {
-    let mut aliases = HashSet::new();
-    let _ = visit_stmts(hir, stmts, |stmt| {
-        let (var_id, value) = match stmt.kind {
-            StmtKind::DeclSingle(var_id) => (Some(var_id), hir.variable(var_id).initializer),
-            StmtKind::Expr(expr) => match &expr.peel_parens().kind {
-                ExprKind::Assign(lhs, _, rhs) => (lhs_local_var(hir, lhs), Some(*rhs)),
-                _ => (None, None),
-            },
-            _ => (None, None),
-        };
-        if let Some(var_id) = var_id
-            && let Some(value) = value
-            && expr_reads_sender(hir, value, &mut HashSet::new(), &aliases)
-        {
-            aliases.insert(var_id);
-        }
-        ControlFlow::Continue(())
-    });
-    aliases
-}
-
-fn expr_reads_sender<'hir>(
-    hir: &'hir hir::Hir<'hir>,
-    expr: &Expr<'_>,
-    seen: &mut HashSet<FunctionId>,
-    aliases: &HashSet<VariableId>,
-) -> bool {
-    expr.visit(&mut |e| {
-        let reads = is_sender_member(e)
-            || underlying_var(e).is_some_and(|v| aliases.contains(&v))
-            || matches!(&e.kind, ExprKind::Call(callee, ..)
-                if function_ids(callee).any(|id| function_reads_sender(hir, id, seen)));
-        if reads { ControlFlow::Break(()) } else { ControlFlow::Continue(()) }
-    })
-    .is_break()
-}
-
-fn function_reads_sender<'hir>(
-    hir: &'hir hir::Hir<'hir>,
-    func_id: FunctionId,
-    seen: &mut HashSet<FunctionId>,
-) -> bool {
-    seen.insert(func_id)
-        && hir.function(func_id).body.is_some_and(|body| {
-            visit_stmts(hir, body.stmts, |stmt| {
-                let reads = stmt_expr(hir, stmt)
-                    .is_some_and(|expr| expr_reads_sender(hir, expr, seen, &HashSet::new()));
-                if reads { ControlFlow::Break(()) } else { ControlFlow::Continue(()) }
-            })
-            .is_break()
-        })
-}
-
-/// State variables read by `expr`, following calls into user functions.
-fn expr_state_vars<'hir>(
-    hir: &'hir hir::Hir<'hir>,
-    expr: &Expr<'_>,
-    seen: &mut HashSet<FunctionId>,
-    out: &mut HashSet<VariableId>,
-) {
-    let _ = expr.visit(&mut |e| {
-        if let Some(var_id) = underlying_var(e)
-            && hir.variable(var_id).kind.is_state()
-        {
-            out.insert(var_id);
-        }
-        if let ExprKind::Call(callee, ..) = &e.kind {
-            for callee_id in function_ids(callee) {
-                function_state_vars(hir, callee_id, seen, out);
-            }
-        }
-        ControlFlow::<()>::Continue(())
-    });
-}
-
-fn function_state_vars<'hir>(
-    hir: &'hir hir::Hir<'hir>,
-    func_id: FunctionId,
-    seen: &mut HashSet<FunctionId>,
-    out: &mut HashSet<VariableId>,
-) {
-    if seen.insert(func_id)
-        && let Some(body) = hir.function(func_id).body
-    {
-        let _ = visit_stmts(hir, body.stmts, |stmt| {
-            if let Some(expr) = stmt_expr(hir, stmt) {
-                expr_state_vars(hir, expr, seen, out);
-            }
-            ControlFlow::Continue(())
-        });
-    }
-}
-
-/// A condition comparing the caller against state.
-fn is_access_check<'hir>(
-    hir: &'hir hir::Hir<'hir>,
-    cond: &Expr<'_>,
-    aliases: &HashSet<VariableId>,
-) -> bool {
-    let mut state_vars = HashSet::new();
-    expr_state_vars(hir, cond, &mut HashSet::new(), &mut state_vars);
-    !state_vars.is_empty() && expr_reads_sender(hir, cond, &mut HashSet::new(), aliases)
-}
-
-/// An `if` whose access-check condition sends one branch into a revert/return.
-fn is_guarding_if<'hir>(
-    hir: &'hir hir::Hir<'hir>,
-    stmt: &'hir Stmt<'hir>,
-    aliases: &HashSet<VariableId>,
-) -> Option<&'hir Expr<'hir>> {
-    let StmtKind::If(cond, then_stmt, else_stmt) = stmt.kind else { return None };
-    (is_access_check(hir, cond, aliases)
-        && (branch_always_exits(then_stmt) || else_stmt.is_some_and(branch_always_exits)))
-    .then_some(cond)
-}
-
-/// True when the function or one of its modifiers contains an access check.
-fn is_protected<'hir>(hir: &'hir hir::Hir<'hir>, func_id: FunctionId) -> bool {
-    modifiers_and_self(hir, func_id).any(|id| has_access_guard(hir, id, &mut HashSet::new()))
-}
-
-fn modifiers_and_self<'hir>(
-    hir: &'hir hir::Hir<'hir>,
-    func_id: FunctionId,
-) -> impl Iterator<Item = FunctionId> + 'hir {
-    hir.function(func_id)
-        .modifiers
-        .iter()
-        .filter_map(|modifier| modifier.id.as_function())
-        .chain(iter::once(func_id))
-}
-
-/// Whether `func_id` (up to its `_` placeholder, if any) checks the caller: a guarding `if`, a
-/// `require`/`assert` on an access check, or a call into a function that does. Bodyless
-/// declarations fall back to a name heuristic.
-fn has_access_guard<'hir>(
-    hir: &'hir hir::Hir<'hir>,
-    func_id: FunctionId,
-    seen: &mut HashSet<FunctionId>,
-) -> bool {
-    if !seen.insert(func_id) {
-        return false;
-    }
-    let func = hir.function(func_id);
-    let Some(body) = func.body else { return looks_like_access_control(func) };
-    let stmts = prefix_stmts(body);
-    let aliases = sender_aliases(hir, stmts);
-    visit_stmts(hir, stmts, |stmt| {
-        let guard = match stmt.kind {
-            StmtKind::If(..) => is_guarding_if(hir, stmt, &aliases).is_some(),
-            _ => stmt_expr(hir, stmt).is_some_and(|expr| {
-                expr.visit(&mut |e| {
-                    let guard = match &e.kind {
-                        ExprKind::Call(callee, args, _) if is_require_or_assert(callee) => {
-                            let cond = args.exprs().next();
-                            cond.is_some_and(|cond| is_access_check(hir, cond, &aliases))
-                        }
-                        ExprKind::Call(callee, ..) => function_ids(callee).any(|id| {
-                            looks_like_access_control(hir.function(id))
-                                || has_access_guard(hir, id, seen)
-                        }),
-                        _ => false,
-                    };
-                    if guard { ControlFlow::Break(()) } else { ControlFlow::Continue(()) }
-                })
-                .is_break()
-            }),
-        };
-        if guard { ControlFlow::Break(()) } else { ControlFlow::Continue(()) }
-    })
-    .is_break()
-}
-
-/// State variables the access checks of `func_id` and its modifiers (up to `_`) depend on.
-fn guard_vars<'hir>(hir: &'hir hir::Hir<'hir>, func_id: FunctionId) -> HashSet<VariableId> {
-    let mut out = HashSet::new();
-    for id in modifiers_and_self(hir, func_id) {
-        let Some(body) = hir.function(id).body else { continue };
-        let stmts = prefix_stmts(body);
-        let aliases = sender_aliases(hir, stmts);
-        let mut seen = HashSet::from([id]);
-        let _ = visit_stmts(hir, stmts, |stmt| {
-            if let Some(cond) = is_guarding_if(hir, stmt, &aliases) {
-                expr_state_vars(hir, cond, &mut seen, &mut out);
-            } else if let Some(expr) = stmt_expr(hir, stmt) {
-                let _ = expr.visit(&mut |e| {
-                    match &e.kind {
-                        ExprKind::Call(callee, args, _) if is_require_or_assert(callee) => {
-                            if let Some(cond) = args.exprs().next()
-                                && is_access_check(hir, cond, &aliases)
-                            {
-                                expr_state_vars(hir, cond, &mut seen, &mut out);
-                            }
-                        }
-                        ExprKind::Call(callee, ..) => {
-                            for callee_id in function_ids(callee) {
-                                if looks_like_access_control(hir.function(callee_id))
-                                    || has_access_guard(hir, callee_id, &mut HashSet::new())
-                                {
-                                    function_state_vars(hir, callee_id, &mut seen, &mut out);
-                                }
-                            }
-                        }
-                        _ => {}
-                    }
-                    ControlFlow::<()>::Continue(())
-                });
-            }
-            ControlFlow::Continue(())
-        });
-    }
-    out
 }
 
 // --- Writes without events --------------------------------------------------------------------
